% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{age_error}
\alias{age_error}
\title{aging error analysis}
\usage{
age_error(
  reader_tester,
  species,
  year,
  area = "goa",
  rec_age = 2,
  plus_age = 45,
  max_age = 100,
  alt = NULL,
  save = TRUE
)
}
\arguments{
\item{species}{= "NORK"}

\item{year}{= year of the assessment}

\item{area}{= "GOA" (BSAI not currently setup)}

\item{rec_age}{= recruitment age}

\item{plus_age}{= max age for modeling}

\item{max_age}{= max age for age error analysis, default 100}

\item{alt}{alternate folder to save to - will be placed in "year/alt/data" folder}

\item{save}{= default is TRUE, FALSE outputs a list, names outputs to a data folder within a specific folder (e.g., save = "age_plus")}

\item{read_tester}{= looks for a file in the user_input folder, e.g., :"reader_tester.csv"}
}
\description{
aging error analysis
}
\examples{
ageage(species = "NORK", year = 2020, admb_home = NULL, area = "GOA", rec_age = 2, plus_age = 45, max_age = 100)
}
