% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{base_plots}
\alias{base_plots}
\title{Base plots}
\usage{
base_plots(year, folder, model_name, rec_age)
}
\arguments{
\item{year}{assessment year}

\item{folder}{the folder with the model in it}

\item{model_name}{the name of your .tpl file}

\item{rec_age}{recruitment age}
}
\description{
Base plots
}
