% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{bts_age_comp}
\alias{bts_age_comp}
\title{trawl survey age comp analysis}
\usage{
bts_age_comp(
  year,
  area = "goa",
  rec_age,
  plus_age,
  rmv_yrs = NULL,
  id = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{area}{default is "goa"}

\item{rec_age}{recruitment age}

\item{plus_age}{plus group age}

\item{rmv_yrs}{any survey years to exclude}

\item{id}{id a specific comp name - will be placed at end of file name e.g., id='use' will create 'bts_age_comp-use.csv' in the data/output folder}

\item{save}{save in the default location}
}
\description{
trawl survey age comp analysis
}
\examples{
bts_age_comp(year = 2020, rec_age = 2, plus_age = 45)
}
