% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{bts_gap_biomass}
\alias{bts_gap_biomass}
\title{GAP survey biomass}
\usage{
bts_gap_biomass(
  year,
  area = "goa",
  type = "region",
  file = NULL,
  rmv_yrs = NULL,
  save = TRUE,
  id = NULL
)
}
\arguments{
\item{year}{model year}

\item{area}{options are bs, bsslope, nbs, ai, goa, old_bs - can only call a single area}

\item{type}{"depth", "stratum", "area", "total", "inpfc", "inpfc_depth" - only available for goa/ai (default: "total") - can only use a single switch}

\item{file}{if not using the design-based abundance, the file name must be stated (e.g. "GAP_VAST.csv") which is stored in "data/user_input"}

\item{rmv_yrs}{any survey years to exclude}

\item{save}{save to default location, default: TRUE}

\item{id}{identifier will be appended to file name - "bts_biomass_vast, default:NULL}
}
\description{
GAP survey biomass
}
