% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{bts_gap_length_comp}
\alias{bts_gap_length_comp}
\title{GAP trawl survey length composition analysis}
\usage{
bts_gap_length_comp(
  year,
  area = "goa",
  lenbins = NULL,
  bysex = NULL,
  rmv_yrs = NULL,
  alt = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{area}{survey area default = "goa"}

\item{lenbins}{lenbin file if left NULL it looks for (year/data/user_input/len_bins.csv")}

\item{bysex}{should the length comp be calculated by sex - default is null (not differentiated)}

\item{rmv_yrs}{any survey years to exclude}

\item{alt}{alternate folder to save to - will be placed in "year/alt/data" folder}

\item{save}{default TRUE}
}
\description{
GAP trawl survey length composition analysis
}
