% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{fac_table}
\alias{fac_table}
\title{Fishery age comp table}
\usage{
fac_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{model being evaluated (folder name)}
}
\description{
Fishery age comp table
}
