% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{fish_length_comp}
\alias{fish_length_comp}
\title{fishery length composition analysis}
\usage{
fish_length_comp(
  year,
  fishery = "fish",
  rec_age,
  lenbins = NULL,
  rmv_yrs = NULL,
  id = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is "fsh1"}

\item{rec_age}{recruitment age}

\item{lenbins}{lenbin file if left NULL it looks for here::here(year, "data", "user_input", "len_bin_labels.csv")}

\item{rmv_yrs}{any survey years to exclude}

\item{id}{id a specific comp name - will be placed at end of file name e.g., id='use' will create 'fsh_length_comp-use.csv' in the data/output folder}

\item{save}{}
}
\description{
fishery length composition analysis
}
