% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{fish_length_comp_pop}
\alias{fish_length_comp_pop}
\title{fishery length composition analysis for pop}
\usage{
fish_length_comp_pop(
  year,
  fishery = "fsh",
  rec_age,
  lenbins = NULL,
  rmv_yrs = NULL,
  alt = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is "fsh"}

\item{rec_age}{recruitment age}

\item{lenbins}{length bins for composition data}

\item{alt}{alternate folder to save to - will be placed in "year/alt/data" folder}

\item{save}{}

\item{fmv_yrs}{years to remove from comp data}
}
\description{
fishery length composition analysis for pop
}
