% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{fsc_table}
\alias{fsc_table}
\title{Fishery size comp table}
\usage{
fsc_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{model being evaluated (folder name)}
}
\description{
Fishery size comp table
}
