% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_tac}
\alias{get_tac}
\title{get last 3 year's TAC values}
\usage{
get_tac(year, area)
}
\arguments{
\item{year}{year of assessment}

\item{area}{"GOA" or "BSAI"}
}
\description{
get last 3 year's TAC values
}
