% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmb.R
\name{pearson}
\alias{pearson}
\title{pearson residuals for RTMB model}
\usage{
pearson(obs, pred, iss, yrs, ind, label = "Age")
}
\arguments{
\item{obs}{observed comp}

\item{pred}{predictedcomp}

\item{iss}{input sample size}

\item{yrs}{comp years}

\item{ind}{vector or ages of lengths}

\item{label}{axis label}
}
\description{
pearson residuals for RTMB model
}
\examples{
\dontrun{
iss = data$fish_age_iss
obs = fish_age_obs
pred = report1$fish_age_pred
yrs = fish_age_yrs
label = "Age"
ind = ages
pearson(obs = fish_age_obs, pred = report1$fish_age_pred, iss = data$fish_age_iss,
    yrs = fish_age_yrs, ind = ages) +  ggtitle('osa fishery age comp residuals')
}
}
