% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot catch}
\usage{
plot_biomass(year, folder, save = TRUE)
}
\arguments{
\item{year}{model year}

\item{folder}{folder name model is in}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Plot catch
}
