% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{plot_compare_biomass}
\alias{plot_compare_biomass}
\title{Plot MCMC estimated biomass}
\usage{
plot_compare_biomass(year, models = c("2020, plus_age", "2020, plus_length"))
}
\arguments{
\item{year}{model year}

\item{models}{year and folder the models are in e.g., c('2020, plus_age', '2020, plus_length')}
}
\description{
Plot MCMC estimated biomass
}
