% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_comps}
\alias{plot_comps}
\title{Plot age and length compositions}
\usage{
plot_comps(year, folder, save = TRUE)
}
\arguments{
\item{year}{= assessment year}

\item{folder}{= folder the model lives in}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Plot age and length compositions
}
\examples{
plot_comps(year, folder)
}
