% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cum_catch}
\alias{plot_cum_catch}
\title{Plot cumulative annual catch}
\usage{
plot_cum_catch(year, save = TRUE)
}
\arguments{
\item{year}{model year}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Plot cumulative annual catch
}
\examples{
\dontrun{
plot_cum_catch(year=2026)
}
}
