% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ofl_abc}
\alias{plot_ofl_abc}
\title{Plot OFL, ABC, TAC and annual catch}
\usage{
plot_ofl_abc(year, save = TRUE)
}
\arguments{
\item{year}{model year}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Plot OFL, ABC, TAC and annual catch
}
\examples{
\dontrun{
plot_ofl_abc(year=2026)
}
}
