% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_params}
\alias{plot_params}
\title{Parameter histogram plots}
\usage{
plot_params(
  year,
  folder,
  model_name,
  pars = c("q_srv1", "ABC", "nattymort", "tot_biom", "F40", "spawn_biom"),
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{folder}{the folder with the model in it}

\item{model_name}{the name of your .tpl file}

\item{pars}{parameter names}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Parameter histogram plots
}
\examples{
plot_swath(year, folder)
}
