% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phase}
\alias{plot_phase}
\title{Phase plot}
\usage{
plot_phase(year, folder, model_name, save = TRUE)
}
\arguments{
\item{year}{model year}

\item{folder}{folder model is in}

\item{model_name}{e.g. goa_nr_2020}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Phase plot
}
