% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_re}
\alias{plot_re}
\title{plot random effects model by GOA area}
\usage{
plot_re(file, save = TRUE)
}
\arguments{
\item{file}{where the files live e.g., "Y:/ABL_MESA/SAFES2020/Apportionment/Plot data - NR"}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
plot random effects model by GOA area
}
\examples{
\dontrun{
plot_re(location)
}

}
