% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rec}
\alias{plot_rec}
\title{Plot recruitment estimates}
\usage{
plot_rec(year, folder, rec_age, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{folder}{folder the model is in}

\item{rec_age}{age at first recruitment}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Plot recruitment estimates
}
