% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rec_ssb}
\alias{plot_rec_ssb}
\title{Recruitment/SSB plot}
\usage{
plot_rec_ssb(year, model, rec_age)
}
\arguments{
\item{year}{model year}

\item{model}{folder model is in}

\item{rec_age}{age at first recruitment}
}
\description{
Recruitment/SSB plot
}
\examples{
plot_rec_ssb(year, model, rec_age)
}
