% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_retro}
\alias{plot_retro}
\title{retrospective plots}
\usage{
plot_retro(year, folder, n_retro = 10, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{folder}{the folder with the model in it}

\item{n_retro}{= 10 number of years to retro}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
retrospective plots
}
\examples{
plot_retro(year, folder)
}
