% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_survey}
\alias{plot_survey}
\title{Plot survey biomass}
\usage{
plot_survey(year, folder, save = TRUE)
}
\arguments{
\item{year}{model year}

\item{folder}{folder model is in}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Plot survey biomass
}
\examples{
plot_survey(year, folder)
}
