% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_swath}
\alias{plot_swath}
\title{Swath plot}
\usage{
plot_swath(year, folder, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{folder}{the folder with the model in it}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Swath plot
}
\examples{
plot_swath(year, folder)
}
