% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{process_results}
\alias{process_results}
\title{Process model results for tables and figs}
\usage{
process_results(
  year,
  folder,
  model_name,
  dat_name,
  rec_age,
  plus_age,
  mcmc = FALSE,
  n_mcmc = 1e+05,
  mcsave = 100,
  len_bins,
  ...
)
}
\arguments{
\item{year}{assessment year}

\item{folder}{model being evaluated (folder name)}

\item{model_name}{name of the folder e.g., updated_nr}

\item{dat_name}{name of dat file e.g., goa_nr_2020}

\item{rec_age}{recruitment age}

\item{plus_age}{plus age group}

\item{mcmc}{was mcmc run? default = FALSE}

\item{n_mcmc}{number of mcmcs run default 100000}

\item{mcsave}{the number of mcmcs saved default 100}

\item{len_bins}{file name (stored in the "user_input" folder) that has length bins}

\item{...}{future functions}
}
\description{
Process model results for tables and figs
}
\examples{
process_results (year = 2020, folder = m18.2, model_name = "goa_nr", dat_name = "goa_nr_2020", rec_age = 2, plus_age = 45, mcmc=FALSE, len_bins = "lbins.csv")

}
