% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmb.R
\name{proj_bio}
\alias{proj_bio}
\title{Title}
\usage{
proj_bio(report, obj = NULL, post = NULL, reps = 500)
}
\arguments{
\item{report}{tmb model report}

\item{obj}{tmb object}

\item{post}{mcmc posterior from tmb}

\item{reps}{how many replicates <= number of mcmc}
}
\value{
data.frame
}
\description{
Title
}
\examples{
\dontrun{
proj_bio(report)
}
}
