% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{saa_waa}
\alias{saa_waa}
\title{size at age analysis}
\usage{
saa_waa(
  year,
  age_data,
  length_data,
  len_bins,
  rec_age,
  alt = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{analysis year}

\item{age_data}{gap survey specimen data}

\item{length_data}{gap survey length data}

\item{rec_age}{recruitment age}

\item{alt}{alternate folder to save to - will be placed in "year/alt/data" folder}

\item{save}{save in the default location}

\item{lenbins}{length bins used e.g., 15:45}
}
\description{
size at age analysis
}
