% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{sac_table}
\alias{sac_table}
\title{Survey age comp table}
\usage{
sac_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{model being evaluated (folder name)}
}
\description{
Survey age comp table
}
