% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_tpl}
\alias{setup_tpl}
\title{Setup folder structure and add .tpl's for a suite of biological functions}
\usage{
setup_tpl(year, dirs = "models")
}
\arguments{
\item{year}{assessment year}

\item{dirs}{directories to write}
}
\value{
creates a designated/named folder structure
}
\description{
Creates a common folder structure for assessment data
}
\examples{
\dontrun{
setup_tpl(2022)
}
}
