% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmb.R
\name{single_like}
\alias{single_like}
\title{Title}
\usage{
single_like(par_name, par_values, obj, fit)
}
\arguments{
\item{par_name}{parameter name}

\item{par_values}{vector of parameter values}

\item{obj}{tmb model object}

\item{fit}{tmb model fit}
}
\value{
likelihood vector
}
\description{
Title
}
\examples{
\dontrun{
logq = seq(log(0.5), log(1.5), length.out=50)
plot(logq,
    single_like(par_name = 'log_q',
                par_values = logq,
                obj=obj2,
                fit = fit2), type = 'l')
}
}
