% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{size_at_age}
\alias{size_at_age}
\title{size at age analysis}
\usage{
size_at_age(
  year,
  area,
  admb_home = NULL,
  rec_age,
  lenbins = NULL,
  alt = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{analysis year}

\item{area}{area data are from e.g., "goa"}

\item{admb_home}{location admb exists on your computer}

\item{rec_age}{recruitment age}

\item{lenbins}{length bin file}

\item{alt}{#' @param alt alternate folder to save to - will be placed in "year/alt/data" folder}

\item{save}{save in the default location}

\item{max_age}{max age for age error analysis - default = 100}
}
\description{
size at age analysis
}
