% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{size_at_age_pop_60}
\alias{size_at_age_pop_60}
\title{size at age for 1960s POP}
\usage{
size_at_age_pop_60(year, rec_age, lenbins = NULL)
}
\arguments{
\item{year}{analysis year}

\item{rec_age}{recruitment age}

\item{lenbins}{length bin file - this should be inyour user_input folder}
}
\description{
size at age for 1960s POP
}
