% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ssc_table}
\alias{ssc_table}
\title{Survey size comp table}
\usage{
ssc_table(year, model)
}
\arguments{
\item{year}{assessment year}

\item{model}{model being evaluated (folder name)}
}
\description{
Survey size comp table
}
