% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_report}
\alias{theme_report}
\title{Set figure theme for reports}
\usage{
theme_report(base_size = 11, base_family = "Times")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\description{
Set figure theme for reports
}
\examples{
\dontrun{
theme_report(base_size = 11, base_family = "Times")

Other fonts are available, though sans font is
the easiest to implement using the following.

theme_report(base_family = "")

Updating font size is accomplished by changing the base_size.

theme_report(base_size = 20, base_family = "")
}
}
