% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{ts_length_comp}
\alias{ts_length_comp}
\title{trawl survey length composition analysis}
\usage{
ts_length_comp(year, area = "goa", lenbins = NULL, bysex = NULL, save = TRUE)
}
\arguments{
\item{year}{assessment year}

\item{area}{survey area default = "goa"}

\item{lenbins}{lenbin file if left NULL it looks for (year/data/user_input/len_bins.csv")}

\item{bysex}{should the length comp be calculated by sex - default is null (not differentiated)}
}
\description{
trawl survey length composition analysis
}
