% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{bts_biomass}
\alias{bts_biomass}
\title{survey biomass}
\usage{
bts_biomass(
  year,
  area = "goa",
  type = "total",
  file = NULL,
  rmv_yrs = NULL,
  alt = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{of interest}

\item{area}{options are bs, bsslope, nbs, ai, goa, old_bs - can only call a single area}

\item{type}{"depth", "stratum", "area", "total", "inpfc", "inpfc_depth" - only available for goa/ai (default: "total") - can only use a single switch}

\item{file}{if not using the design-based abundance, the file name must be stated (e.g. "GAP_VAST.csv")}

\item{rmv_yrs}{any survey years to exclude}

\item{alt}{alternate folder to save to - will be placed in "year/alt/data" folder}

\item{save}{save to default location}
}
\description{
survey biomass
}
