% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{bts_length_comp}
\alias{bts_length_comp}
\title{trawl survey length composition analysis}
\usage{
bts_length_comp(
  year,
  area = "goa",
  sa_index,
  lenbins = NULL,
  bysex = NULL,
  alt = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{area}{survey area default = "goa"}

\item{sa_index}{index for size-age transition matrix}

\item{lenbins}{lenbin file if left NULL it looks for (year/data/user_input/len_bins.csv")}

\item{bysex}{should the length comp be calculated by sex - default is null (not differentiated)}

\item{alt}{alternate folder to save to - will be placed in "year/alt/data" folder}
}
\description{
trawl survey length composition analysis
}
