% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{fish_age_comp}
\alias{fish_age_comp}
\title{fishery age composition analysis}
\usage{
fish_age_comp(
  year,
  fishery = "fish",
  rec_age,
  plus_age,
  rmv_yrs = NULL,
  id = NULL,
  save = TRUE
)
}
\arguments{
\item{year}{assessment year}

\item{fishery}{default is fsh, change if age comps from multiple fisheries (e.g., fsh1, fsh2)}

\item{rec_age}{recruitment age}

\item{plus_age}{plus age group}

\item{rmv_yrs}{any years to remove form the age comp e.g. c(1987, 1989)}

\item{id}{id a specific comp name - will be placed at end of file name e.g., id='use' will create 'fsh_age_comp-use.csv' in the data/output folder}

\item{save}{whether to save the file - wll be placed in "year/data/output" folder}
}
\description{
fishery age composition analysis
}
\examples{
\dontrun{
fish_age_comp(year, fishery = "fsh", rec_age, plus_age)
}
}
