% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rec}
\alias{plot_rec}
\title{Plot recruitment estimates}
\usage{
plot_rec(year, model, rec_age)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}
}
\description{
Plot recruitment estimates
}
