% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_selex}
\alias{plot_selex}
\title{Plot selectivity at age}
\usage{
plot_selex(year, folder, save = TRUE)
}
\arguments{
\item{year}{model year}

\item{folder}{folder model is in}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Plot selectivity at age
}
\examples{
plot_selex(year, folder)
}
