% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_ak}
\alias{proj_ak}
\title{Understanding proj (or not)}
\usage{
proj_ak(
  year,
  last_full_assess,
  alt = NULL,
  folder,
  species,
  region,
  rec_age = 2,
  off_yr = FALSE,
  run_name = "Standard",
  tac_abc = 1,
  srr = 1,
  rec_proj = 1,
  srr_cond = 0,
  srr_prior = 0,
  nyrs_proj = 14,
  nsims = 1000,
  n_species = 1,
  abc_mult = 1,
  pop_scalar = 1000,
  tac_categories = 1,
  tac_models = 1,
  admb_home = NULL
)
}
\arguments{
\item{year}{model year}

\item{last_full_assess}{year that the last full assessment was done.}

\item{alt}{prefix folder structure e.g. "mgmt" - default = NULL folder structure: 'year\alt'}

\item{folder}{folder that model is in, folder structure: 'year\alt\folder'}

\item{species}{e.g., "dusky"}

\item{region}{e.g., "goa"}

\item{rec_age}{recruitment age, default: 2}

\item{off_yr}{is this a full assessment or off year, default: FALSE}

\item{run_name}{- default: "Standard"}

\item{tac_abc}{-  set TAC equal to ABC (1 means true, otherwise false)",}

\item{srr}{- Stock-recruitment type (1=Ricker, 2=Bholt)}

\item{rec_proj}{projection rec form (default: 1 = use observed mean and std, option 2 = use estimated SRR and estimated sigma R)}

\item{srr_cond}{SR-Conditioning (0 means no, 1 means use Fmsy == F35%?, 2 means Fmsy == F35% and Bmsy=B35%  condition (affects SRR fits)}

\item{srr_prior}{a prior that mean historical recruitment is similar to expected recruitment at half mean SSB and double mean SSB 0 means don't use, otherwise specify CV}

\item{nyrs_proj}{Number of projection years default: 14}

\item{nsims}{Number of simulations default: 1000}

\item{n_species}{number of species being evaluated: default: 1}

\item{abc_mult}{abc multiplier, default: 1}

\item{pop_scalar}{population scalar, default: 1000}

\item{tac_categories}{default 1}

\item{tac_models}{default 1}

\item{admb_home}{location of ADMB if different than c:/admb}

\item{model_name}{name of the assessment model e.g., "m18.2" not the projection model, this is hopefully the same name as the folder...}
}
\description{
workflow:
  1. setup folders
  2. populate folders with "built in files"
  3. populate folders with assessment info
  4. run model
      a. max F output
      b. author F output
  5. compile and retrieve results
}
