% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{write_ctl_pop}
\alias{write_ctl_pop}
\title{write a .ctl file for goa pop}
\usage{
write_ctl_pop(
  year,
  base_mdl_fldr,
  curr_mdl_fldr,
  mdl_name = "Model_1",
  ctl_name = "goa_pop",
  dat_name = "goa_pop",
  folder,
  mcmc = FALSE
)
}
\arguments{
\item{year}{assessment year}

\item{base_mdl_fldr}{name of folder in which last full assessment model is stored (should be in 'models' folder)}

\item{curr_mdl_fldr}{name of folder that stores current model}

\item{mdl_name}{optional name for model run (default = "Model_1)}

\item{ctl_name}{what you want to name your ctl file}

\item{dat_name}{name of data file}

\item{folder}{where you want ctl file written}

\item{mcmc}{toggle to freeze sigr in mcmc estimation}
}
\description{
write a .ctl file for goa pop
}
